Removed broken paging. Begin upgrade to 64 bit.

+++ kernel/vga.c
@@ -95,7 +95,7 @@ void vprint(char *str) {
 	}
 }
 
-void vputn(u32 n) {
+void vputn(u64 n) {
 	if (n == 0) {
 		vputc('0');
 		return;
